fix signed vs. unsigned comparison warning in GL_TextureMode_f
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@9995 d7cf8633-e32d-0410-b094-e92efae38249
::stable-branch::merge=d032cfb5ca1e47238f2bd96886793c21547f0605

+++ gl_textures.c
@@ -338,7 +338,7 @@ static void GL_TextureMode_f (void)
 		return;
 	}
 
-	for (i = 0;i < sizeof(modes)/sizeof(*modes);i++)
+	for (i = 0;i < (int)(sizeof(modes)/sizeof(*modes));i++)
 		if (!strcasecmp (modes[i].name, Cmd_Argv(1) ) )
 			break;
 	if (i == 6)
